# perf(REST): RHICOMPL-2249 use a subquery for RuleResults by hosts