--- a/app/policies/rule_result_policy.rb
+++ b/app/policies/rule_result_policy.rb
@@ -13,7 +13,7 @@ class RuleResultPolicy < ApplicationPolicy
   # Only show RuleResults belonging to hosts visible by the current user
   class Scope < ::ApplicationPolicy::Scope
     def resolve
-      available_hosts = HostPolicy::Scope.new(user, Host).resolve.pluck(:id)
+      available_hosts = HostPolicy::Scope.new(user, Host).resolve.select(:id)
       scope.where(host_id: available_hosts)
     end
   end