:runner: :racehorse: :zap: 

**Before:**
Completed 200 OK in 13276ms (Views: 3579.5ms | ActiveRecord: 3986.8ms)
**After:**
Completed 200 OK in 5632ms (Views: 56.7ms | ActiveRecord: 5537.0ms)